requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-telegram-bot==20.6
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.21
alembic==1.12.0
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
